test: Added unit tests for BookService with JUnit and Mockito, then deleteBook change to fit into it

+++ src/main/java/com/davi/Bookshelf/services/BookService.java
@@ -53,6 +53,9 @@ public class BookService {
     }
 
     public void deleteBook(UUID id){
+        if (!bookRepository.existsById(id))
+            throw new ResourceNotFoundException("Livro não encontrado");
+
         bookRepository.deleteById(id);
     }
 }
